Version 0.0.2

+++ extraction/extraction_entry.sh
@@ -4,15 +4,24 @@
 dir_path="/home/ubuntu/data/raw/linkedln"
 
 #Trigger the extraction script to pull the job posting and save it as json in the local environment
-python3 /home/ubuntu/development/batch_linkedln_jobs/extraction/linkedln_job_scraper.py --result_path_folder $dir_path --pull_limit 5
+python3 /home/ubuntu/development/batch_linkedln_jobs/extraction/linkedln_job_scraper.py --result_path_folder $dir_path --pull_limit 200
 echo "Extraction is complete and result is saved to $dir_path"
 
 #Trigger the python script to upload the result json to S3
 for i in `ls $dir_path`
 do
-    python3 /home/ubuntu/development/batch_linkedln_jobs/extraction/upload_to_s3.py --local_path $dir_path/$i --bucket linkedln-jobs-etl --s3path raw/$i
-    echo "$i is successfully uploaded to S3" 
-    rm $dir_path/$i
-    echo "$i is successfully removed after uploading to S3" 
+    if [ -s $dir_path/$i ]; then 
+        # The file is not-empty.
+        python3 /home/ubuntu/development/batch_linkedln_jobs/extraction/upload_to_s3.py --local_path $dir_path/$i --bucket linkedln-jobs-etl --s3path raw/$i
+        echo "$i is successfully uploaded to S3" 
+        rm $dir_path/$i
+        echo "$i is successfully removed after uploading to S3" 
+    else
+        # The file is empty.
+        echo "$i is empty, no action is taken"
+        exit
+    fi
+
 done
+
